исправлен вывод без создания лишней переменной

+++ zakozhurnikova.kristina/S2/outputResult.cpp
@@ -3,16 +3,16 @@
 
 void zakozhurnikova::outputResult(std::ostream& out, Stack< ll >& result)
 {
-  unsigned long long quantity = 0;
-  while (!result.empty())
+  if (!result.empty())
   {
-    if (quantity != 0)
-    {
-      out << " ";
-    }
     out << result.top();
     result.drop();
-    ++quantity;
+  }
+  while (!result.empty())
+  {
+    out << " ";
+    out << result.top();
+    result.drop();
   }
   out << '\n';
 }
